backend/requirements.txt
flask
flask-cors
python-dotenv
requests
pyjwt
stripe
pyodbc
pydantic